Ismoil

+++ TeamProjectMVC/Entity/Product.cs
@@ -6,7 +6,7 @@
 
         public string? Name { get; set; }
 
-        public int Quantity { get; set; }
+        public ulong Quantity { get; set; }
 
         public double Price { get; set; }
