+++ TeamProjectMVC/Repository/IProductRepository.cs
@@ -10,7 +10,7 @@ namespace TeamProjectMVC.Repository
         Task Add(string userId, string userName, ProductDTO productDto);
         Task Update(string userId , string userName, string id, ProductDTO productDto);
         Task Delete(string userId, string userName, string id);
-        Task<double> CalculateTotalPrice(int quantity, double price);
+        Task<double> CalculateTotalPrice(ulong quantity, double price);
         Task<RoleProductDTO> RetrieveDto(string userId, string userName, string role);
     }
 }
